+++ beatmap/v2/obstacle.ts
@@ -11,8 +11,8 @@ export class Obstacle extends WrapObstacle<IObstacle> {
       _time: 0,
       _lineIndex: 0,
       _type: 0,
-      _duration: 1,
-      _width: 1,
+      _duration: 0,
+      _width: 0,
       _customData: {},
    };