+++ beatmap/v3/basicEvent.ts
@@ -16,7 +16,7 @@ export class BasicEvent extends WrapEvent<IBasicEvent> {
       b: 0,
       et: 0,
       i: 0,
-      f: 1,
+      f: 0,
       customData: {},
    };