+++ beatmap/v3/clean.ts
@@ -1,14 +1,7 @@
 import { round } from '../../utils/math.ts';
 import { IDifficulty } from '../../types/beatmap/v3/difficulty.ts';
 import { ICleanOptions } from '../../types/beatmap/shared/clean.ts';
-import { deepClean } from '../shared/clean.ts';
-
-// deno-lint-ignore no-explicit-any
-function purgeZeros(data: Record<string, any>) {
-   for (const k in data) {
-      if (typeof data[k] === 'number' && data[k] === 0) delete data[k];
-   }
-}
+import { deepClean, purgeZeros } from '../shared/clean.ts';
 
 export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
    for (const i1 in data.bpmEvents) {
@@ -17,7 +10,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.rotationEvents) {
       const o1 = data.rotationEvents[i1];
@@ -29,7 +22,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.colorNotes) {
       const o1 = data.colorNotes[i1];
@@ -40,7 +33,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.bombNotes) {
       const o1 = data.bombNotes[i1];
@@ -51,7 +44,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.obstacles) {
       const o1 = data.obstacles[i1];
@@ -63,7 +56,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.sliders) {
       const o1 = data.sliders[i1];
@@ -77,7 +70,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.burstSliders) {
       const o1 = data.burstSliders[i1];
@@ -90,7 +83,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.waypoints) {
       const o1 = data.waypoints[i1];
@@ -101,7 +94,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.basicBeatmapEvents) {
       const o1 = data.basicBeatmapEvents[i1];
@@ -113,7 +106,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.colorBoostBeatmapEvents) {
       const o1 = data.colorBoostBeatmapEvents[i1];
@@ -125,7 +118,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
          delete o1.customData;
       }
       if (!o1.o) delete (o1 as Partial<typeof o1>).o;
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.lightColorEventBoxGroups) {
       const o1 = data.lightColorEventBoxGroups[i1];
@@ -153,7 +146,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
             if (!Object.keys(o3.customData!).length) {
                delete o3.customData;
             }
-            purgeZeros(o3);
+            if (options.purgeZeros) purgeZeros(o3);
          }
          deepClean(
             o2.f.customData!,
@@ -163,7 +156,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
          if (!Object.keys(o2.f.customData!).length) {
             delete o2.f.customData;
          }
-         purgeZeros(o2.f);
+         if (options.purgeZeros) purgeZeros(o2.f);
          deepClean(
             o2.customData!,
             `difficulty.lightColorEventBoxGroups[${i1}].e[${i2}].customData`,
@@ -172,13 +165,13 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
          if (!Object.keys(o2.customData!).length) {
             delete o2.customData;
          }
-         purgeZeros(o2);
+         if (options.purgeZeros) purgeZeros(o2);
       }
       deepClean(o1.customData!, `difficulty.lightColorEventBoxGroups[${i1}].customData`, options);
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.lightRotationEventBoxGroups) {
       const o1 = data.lightRotationEventBoxGroups[i1];
@@ -206,7 +199,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
             if (!Object.keys(o3.customData!).length) {
                delete o3.customData;
             }
-            purgeZeros(o3);
+            if (options.purgeZeros) purgeZeros(o3);
          }
          deepClean(
             o2.f.customData!,
@@ -216,7 +209,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
          if (!Object.keys(o2.f.customData!).length) {
             delete o2.f.customData;
          }
-         purgeZeros(o2.f);
+         if (options.purgeZeros) purgeZeros(o2.f);
          deepClean(
             o2.customData!,
             `difficulty.lightRotationEventBoxGroups[${i2}].e[${i2}].customData`,
@@ -225,7 +218,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
          if (!Object.keys(o2.customData!).length) {
             delete o2.customData;
          }
-         purgeZeros(o2);
+         if (options.purgeZeros) purgeZeros(o2);
       }
       deepClean(
          o1.customData!,
@@ -235,7 +228,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.lightTranslationEventBoxGroups) {
       const o1 = data.lightTranslationEventBoxGroups[i1];
@@ -263,7 +256,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
             if (!Object.keys(o3.customData!).length) {
                delete o3.customData;
             }
-            purgeZeros(o3);
+            if (options.purgeZeros) purgeZeros(o3);
          }
          deepClean(
             o2.f.customData!,
@@ -273,7 +266,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
          if (!Object.keys(o2.f.customData!).length) {
             delete o2.f.customData;
          }
-         purgeZeros(o2.f);
+         if (options.purgeZeros) purgeZeros(o2.f);
          deepClean(
             o2.customData!,
             `difficulty.lightTranslationEventBoxGroups[${i1}].e[${i2}].customData`,
@@ -282,7 +275,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
          if (!Object.keys(o2.customData!).length) {
             delete o2.customData;
          }
-         purgeZeros(o2);
+         if (options.purgeZeros) purgeZeros(o2);
       }
       deepClean(
          o1.customData!,
@@ -292,7 +285,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data.vfxEventBoxGroups) {
       const o1 = data.vfxEventBoxGroups[i1];
@@ -314,7 +307,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
          if (!Object.keys(o2.f.customData!).length) {
             delete o2.f.customData;
          }
-         purgeZeros(o2.f);
+         if (options.purgeZeros) purgeZeros(o2.f);
          deepClean(
             o2.customData!,
             `difficulty.vfxEventBoxGroups[${i1}].e[${i2}].customData`,
@@ -323,13 +316,13 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
          if (!Object.keys(o2.customData!).length) {
             delete o2.customData;
          }
-         purgeZeros(o2);
+         if (options.purgeZeros) purgeZeros(o2);
       }
       deepClean(o1.customData!, `difficulty.vfxEventBoxGroups[${i1}].customData`, options);
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data._fxEventsCollection._fl) {
       const o1 = data._fxEventsCollection._fl[i1];
@@ -341,7 +334,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const i1 in data._fxEventsCollection._il) {
       const o1 = data._fxEventsCollection._il[i1];
@@ -352,7 +345,7 @@ export function cleanDifficulty(data: IDifficulty, options: ICleanOptions) {
       if (!Object.keys(o1.customData!).length) {
          delete o1.customData;
       }
-      purgeZeros(o1);
+      if (options.purgeZeros) purgeZeros(o1);
    }
    for (const o1 of data.basicEventTypesWithKeywords.d) {
       if (options.stringTrim) {