+++ types/beatmap/shared/clean.ts
@@ -3,6 +3,8 @@ export interface ICleanOptions {
    floatTrim?: number;
    /** Trim string. */
    stringTrim?: boolean;
+   /** Remove zero-valued attribute in JSON. */
+   purgeZeros?: boolean;
    /** Throw error when encountering null or undefined value. */
    throwError?: boolean;
 }